+++ cmd/history.go
@@ -15,6 +15,7 @@
 package cmd
 
 import (
+	"encoding/json"
 	"fmt"
 	"sort"
 	"strings"
@@ -123,6 +124,14 @@ func displayUpdatesJSON(updates []backend.UpdateInfo, decrypter config.Decrypter
 				value, err := v.Value(decrypter)
 				contract.AssertNoError(err)
 				configValue.Value = makeStringRef(value)
+
+				if v.Object() {
+					var obj interface{}
+					if err := json.Unmarshal([]byte(value), &obj); err != nil {
+						return err
+					}
+					configValue.ObjectValue = obj
+				}
 			}
 
 			info.Config[k.String()] = configValue