+++ cmd/watch.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2018, Pulumi Corporation.
+// Copyright 2016-2019, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
 package cmd
 
 import (
+	"context"
+
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -25,9 +27,10 @@ import (
 	"github.com/pulumi/pulumi/pkg/util/result"
 )
 
-func newPreviewCmd() *cobra.Command {
+// intentionally disabling here for cleaner err declaration/assignment.
+// nolint: vetshadow
+func newWatchCmd() *cobra.Command {
 	var debug bool
-	var expectNop bool
 	var message string
 	var stack string
 	var configArray []string
@@ -35,59 +38,42 @@ func newPreviewCmd() *cobra.Command {
 
 	// Flags for engine.UpdateOptions.
 	var policyPackPaths []string
-	var diffDisplay bool
-	var eventLogPath string
-	var jsonDisplay bool
 	var parallel int
+	var refresh bool
 	var showConfig bool
 	var showReplacementSteps bool
 	var showSames bool
-	var showReads bool
-	var suppressOutputs bool
+	var secretsProvider string
 
 	var cmd = &cobra.Command{
-		Use:        "preview",
-		Aliases:    []string{"pre"},
-		SuggestFor: []string{"build", "plan"},
-		Short:      "Show a preview of updates to a stack's resources",
-		Long: "Show a preview of updates a stack's resources.\n" +
+		Use:        "watch",
+		SuggestFor: []string{"developer", "dev"},
+		Short:      "Continuously update the resources in a stack",
+		Long: "Continuously update the resources in a stack.\n" +
 			"\n" +
-			"This command displays a preview of the updates to an existing stack whose state is\n" +
-			"represented by an existing state file. The new desired state is computed by running\n" +
-			"a Pulumi program, and extracting all resource allocations from its resulting object graph.\n" +
-			"These allocations are then compared against the existing state to determine what\n" +
-			"operations must take place to achieve the desired state. No changes to the stack will\n" +
-			"actually take place.\n" +
+			"This command watches the working directory for the current project and updates the active stack whenever\n" +
+			"the project changes.  In parallel, logs are collected for all resources in the stack and displayed along\n" +
+			"with update progress.\n" +
 			"\n" +
-			"The program to run is loaded from the project in the current directory. Use the `-C` or\n" +
+			"The program to watch is loaded from the project in the current directory by default. Use the `-C` or\n" +
 			"`--cwd` flag to use a different directory.",
-		Args: cmdutil.NoArgs,
+		Args: cmdutil.MaximumNArgs(1),
 		Run: cmdutil.RunResultFunc(func(cmd *cobra.Command, args []string) result.Result {
-			var displayType = display.DisplayProgress
-			if diffDisplay {
-				displayType = display.DisplayDiff
+
+			opts, err := updateFlagsToOptions(false /* interactive */, true /* skippreview*/, true /* autoapprove*/)
+			if err != nil {
+				return result.FromError(err)
 			}
 
-			opts := backend.UpdateOptions{
-				Engine: engine.UpdateOptions{
-					LocalPolicyPackPaths: policyPackPaths,
-					Parallel:             parallel,
-					Debug:                debug,
-					UseLegacyDiff:        useLegacyDiff(),
-				},
-				Display: display.Options{
-					Color:                cmdutil.GetGlobalColorization(),
-					ShowConfig:           showConfig,
-					ShowReplacementSteps: showReplacementSteps,
-					ShowSameResources:    showSames,
-					ShowReads:            showReads,
-					SuppressOutputs:      suppressOutputs,
-					IsInteractive:        cmdutil.Interactive(),
-					Type:                 displayType,
-					JSONDisplay:          jsonDisplay,
-					EventLogPath:         eventLogPath,
-					Debug:                debug,
-				},
+			opts.Display = display.Options{
+				Color:                cmdutil.GetGlobalColorization(),
+				ShowConfig:           showConfig,
+				ShowReplacementSteps: showReplacementSteps,
+				ShowSameResources:    showSames,
+				SuppressOutputs:      true,
+				IsInteractive:        false,
+				Type:                 display.DisplayWatch,
+				Debug:                debug,
 			}
 
 			s, err := requireStack(stack, true, opts.Display, true /*setCurrent*/)
@@ -105,7 +91,7 @@ func newPreviewCmd() *cobra.Command {
 				return result.FromError(err)
 			}
 
-			m, err := getUpdateMetadata("", root)
+			m, err := getUpdateMetadata(message, root)
 			if err != nil {
 				return result.FromError(errors.Wrap(err, "gathering environment metadata"))
 			}
@@ -120,7 +106,15 @@ func newPreviewCmd() *cobra.Command {
 				return result.FromError(errors.Wrap(err, "getting stack configuration"))
 			}
 
-			changes, res := s.Preview(commandContext(), backend.UpdateOperation{
+			opts.Engine = engine.UpdateOptions{
+				LocalPolicyPackPaths: policyPackPaths,
+				Parallel:             parallel,
+				Debug:                debug,
+				Refresh:              refresh,
+				UseLegacyDiff:        useLegacyDiff(),
+			}
+
+			res := s.Watch(commandContext(), backend.UpdateOperation{
 				Proj:               proj,
 				Root:               root,
 				M:                  m,
@@ -129,12 +123,11 @@ func newPreviewCmd() *cobra.Command {
 				SecretsManager:     sm,
 				Scopes:             cancellationScopes,
 			})
-
 			switch {
+			case res != nil && res.Error() == context.Canceled:
+				return result.FromError(errors.New("update cancelled"))
 			case res != nil:
 				return PrintEngineResult(res)
-			case expectNop && changes != nil && changes.HasChanges():
-				return result.FromError(errors.New("error: no changes were expected but changes were proposed"))
 			default:
 				return nil
 			}
@@ -144,9 +137,6 @@ func newPreviewCmd() *cobra.Command {
 	cmd.PersistentFlags().BoolVarP(
 		&debug, "debug", "d", false,
 		"Print detailed debugging output during resource operations")
-	cmd.PersistentFlags().BoolVar(
-		&expectNop, "expect-no-changes", false,
-		"Return an error if any changes are proposed by this preview")
 	cmd.PersistentFlags().StringVarP(
 		&stack, "stack", "s", "",
 		"The name of the stack to operate on. Defaults to the current stack")
@@ -155,52 +145,40 @@ func newPreviewCmd() *cobra.Command {
 		"Use the configuration values in the specified file rather than detecting the file name")
 	cmd.PersistentFlags().StringArrayVarP(
 		&configArray, "config", "c", []string{},
-		"Config to use during the preview")
+		"Config to use during the update")
 	cmd.PersistentFlags().BoolVar(
 		&configPath, "config-path", false,
 		"Config keys contain a path to a property in a map or list to set")
+	cmd.PersistentFlags().StringVar(
+		&secretsProvider, "secrets-provider", "default", "The type of the provider that should be used to encrypt and "+
+			"decrypt secrets (possible choices: default, passphrase, awskms, azurekeyvault, gcpkms, hashivault). Only"+
+			"used when creating a new stack from an existing template")
 
 	cmd.PersistentFlags().StringVarP(
 		&message, "message", "m", "",
-		"Optional message to associate with the preview operation")
+		"Optional message to associate with each update operation")
 
 	// Flags for engine.UpdateOptions.
 	if hasDebugCommands() {
 		cmd.PersistentFlags().StringSliceVar(
 			&policyPackPaths, "policy-pack", []string{},
-			"Run one or more analyzers as part of this update")
+			"Run one or more policy packs as part of each update")
 	}
-	cmd.PersistentFlags().BoolVar(
-		&diffDisplay, "diff", false,
-		"Display operation as a rich diff showing the overall change")
-	cmd.Flags().BoolVarP(
-		&jsonDisplay, "json", "j", false,
-		"Serialize the preview diffs, operations, and overall output as JSON")
 	cmd.PersistentFlags().IntVarP(
 		&parallel, "parallel", "p", defaultParallel,
 		"Allow P resource operations to run in parallel at once (1 for no parallelism). Defaults to unbounded.")
+	cmd.PersistentFlags().BoolVarP(
+		&refresh, "refresh", "r", false,
+		"Refresh the state of the stack's resources before each update")
 	cmd.PersistentFlags().BoolVar(
 		&showConfig, "show-config", false,
 		"Show configuration keys and variables")
 	cmd.PersistentFlags().BoolVar(
 		&showReplacementSteps, "show-replacement-steps", false,
 		"Show detailed resource replacement creates and deletes instead of a single step")
-
 	cmd.PersistentFlags().BoolVar(
 		&showSames, "show-sames", false,
-		"Show resources that needn't be updated because they haven't changed, alongside those that do")
-	cmd.PersistentFlags().BoolVar(
-		&showReads, "show-reads", false,
-		"Show resources that are being read in, alongside those being managed directly in the stack")
+		"Show resources that don't need be updated because they haven't changed, alongside those that do")
 
-	cmd.PersistentFlags().BoolVar(
-		&suppressOutputs, "suppress-outputs", false,
-		"Suppress display of stack outputs (in case they contain sensitive values)")
-
-	if hasDebugCommands() {
-		cmd.PersistentFlags().StringVar(
-			&eventLogPath, "event-log", "",
-			"Log events to a file at this path")
-	}
 	return cmd
 }