+++ pkg/backend/httpstate/stack.go
@@ -151,6 +151,10 @@ func (s *cloudStack) Destroy(ctx context.Context, op backend.UpdateOperation) (e
 	return backend.DestroyStack(ctx, s, op)
 }
 
+func (s *cloudStack) Watch(ctx context.Context, op backend.UpdateOperation) result.Result {
+	return backend.WatchStack(ctx, s, op)
+}
+
 func (s *cloudStack) GetLogs(ctx context.Context, cfg backend.StackConfiguration,
 	query operations.LogQuery) ([]operations.LogEntry, error) {
 	return backend.GetStackLogs(ctx, s, cfg, query)