+++ pkg/backend/stack.go
@@ -47,6 +47,8 @@ type Stack interface {
 	Refresh(ctx context.Context, op UpdateOperation) (engine.ResourceChanges, result.Result)
 	// Destroy this stack's resources.
 	Destroy(ctx context.Context, op UpdateOperation) (engine.ResourceChanges, result.Result)
+	// Watch this stack.
+	Watch(ctx context.Context, op UpdateOperation) result.Result
 
 	// remove this stack.
 	Remove(ctx context.Context, force bool) (bool, error)
@@ -88,6 +90,12 @@ func RefreshStack(ctx context.Context, s Stack, op UpdateOperation) (engine.Reso
 // DestroyStack destroys all of this stack's resources.
 func DestroyStack(ctx context.Context, s Stack, op UpdateOperation) (engine.ResourceChanges, result.Result) {
 	return s.Backend().Destroy(ctx, s, op)
+}
+
+// WatchStack watches the projects working directory for changes and automatically updates the
+// active stack.
+func WatchStack(ctx context.Context, s Stack, op UpdateOperation) result.Result {
+	return s.Backend().Watch(ctx, s, op)
 }
 
 // GetLatestConfiguration returns the configuration for the most recent deployment of the stack.