+++ scripts/make_release.sh
@@ -49,6 +49,7 @@ copy_package() {
 run_go_build "${ROOT}"
 run_go_build "${ROOT}/sdk/nodejs/cmd/pulumi-language-nodejs"
 run_go_build "${ROOT}/sdk/python/cmd/pulumi-language-python"
+run_go_build "${ROOT}/sdk/dotnet/cmd/pulumi-language-dotnet"
 run_go_build "${ROOT}/sdk/go/pulumi-language-go"
 
 # Copy over the language and dynamic resource providers.