+++ sdk/nodejs/runtime/invoke.ts
@@ -12,8 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+const asyncify = require("callback-to-async-iterator");
+import { AsyncIterable } from "@pulumi/query/interfaces";
 import * as fs from "fs";
 import * as grpc from "grpc";
+
 import * as asset from "../asset";
 import { InvokeOptions } from "../invoke";
 import * as log from "../log";
@@ -76,6 +79,61 @@ export function invoke(tok: string, props: Inputs, opts: InvokeOptions = {}): Pr
     }
 
     return invokeSync(tok, props, opts, syncInvokes);
+}
+
+export async function streamInvoke(
+    tok: string,
+    props: Inputs,
+    opts: InvokeOptions = {},
+): Promise<StreamInvokeResponse<any>> {
+    const label = `StreamInvoking function: tok=${tok} asynchronously`;
+    log.debug(label + (excessiveDebugOutput ? `, props=${JSON.stringify(props)}` : ``));
+
+    // Wait for all values to be available, and then perform the RPC.
+    const done = rpcKeepAlive();
+    try {
+        const serialized = await serializeProperties(`streamInvoke:${tok}`, props);
+        log.debug(
+            `StreamInvoke RPC prepared: tok=${tok}` + excessiveDebugOutput
+                ? `, obj=${JSON.stringify(serialized)}`
+                : ``,
+        );
+
+        // Fetch the monitor and make an RPC request.
+        const monitor: any = getMonitor();
+
+        const provider = await ProviderResource.register(getProvider(tok, opts));
+        const req = createInvokeRequest(tok, serialized, provider, opts);
+
+        // Call `streamInvoke`.
+        const call = monitor.streamInvoke(req, {});
+
+        // Transform the callback-oriented `streamInvoke` result into a plain-old (potentially
+        // infinite) `AsyncIterable`.
+        const listenForWatchEvents = (callback: (obj: any) => void) => {
+            return new Promise(resolve => {
+                call.on("data", function(thing: any) {
+                    const live = deserializeResponse(tok, thing);
+                    callback(live);
+                });
+                call.on("error", (err: any) => {
+                    if (err.code === 1 && err.details === "Cancelled") {
+                        return;
+                    }
+                    throw err;
+                });
+                // Infinite stream, never call `resolve`.
+            });
+        };
+        const stream: AsyncIterable<any> = asyncify.default(listenForWatchEvents);
+
+        // Return a cancellable handle to the stream.
+        return new StreamInvokeResponse(
+            stream,
+            () => call.cancel());
+    } finally {
+        done();
+    }
 }
 
 export function invokeFallbackToAsync(tok: string, props: Inputs, opts: InvokeOptions): Promise<any> {
@@ -178,6 +236,24 @@ For more details see: https://www.pulumi.com/docs/troubleshooting/#synchronous-c
     }
 }
 
+// StreamInvokeResponse represents a (potentially infinite) streaming response to `streamInvoke`,
+// with facilities to gracefully cancel and clean up the stream.
+export class StreamInvokeResponse<T> implements AsyncIterable<T> {
+    constructor(
+        private source: AsyncIterable<T>,
+        private cancelSource: () => void,
+    ) {}
+
+    // cancel signals the `streamInvoke` should be cancelled and cleaned up gracefully.
+    public cancel() {
+        this.cancelSource();
+    }
+
+    [Symbol.asyncIterator]() {
+        return this.source[Symbol.asyncIterator]();
+    }
+}
+
 // Expose the properties of the actual result of invoke directly on the promise itself. Note this
 // doesn't actually involve any asynchrony.  The promise will be created synchronously and the
 // values copied to it can be used immediately.  We simply make a Promise so that any consumers that
@@ -259,7 +335,7 @@ function serializePropertiesSync(prop: any): any {
     }
 }
 
-function deserializeResponse(tok: string, resp: any) {
+function deserializeResponse(tok: string, resp: any): any {
     const failures: any = resp.getFailuresList();
     if (failures && failures.length) {
         let reasons = "";
@@ -274,5 +350,8 @@ function deserializeResponse(tok: string, resp: any) {
         throw new Error(`Invoke of '${tok}' failed: ${reasons}`);
     }
 
-    return deserializeProperties(resp.getReturn());
+    const ret = resp.getReturn();
+    return ret === undefined
+        ? ret
+        : deserializeProperties(ret);
 }