+++ CHANGELOG.md
@@ -3,6 +3,29 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+- `pulumi policy publish` now determines the Policy Pack name from the Policy Pack, and the
+  the `org-name` CLI argument is now optional. If not specified; the current user account is
+  used.
+  [#3459](https://github.com/pulumi/pulumi/pull/3459)
+
+## 1.5.0 (2019-11-06)
+
+- Include the .NET language provider in the Windows SDK.
+
+## 1.5.0 (2019-11-06)
+
+- Gracefully handle errors when resources use duplicate aliases.
+
+- Use the update token for renew_lease calls and update the API version to 5.
+  [#3348](https://github.com/pulumi/pulumi/pull/3348)
+
+- Improve startup time performance by 0.5-1s by checking for a newer CLI release in parallel.
+  [#3441](https://github.com/pulumi/pulumi/pull/3441)
+
+- Add an experimental `pulumi watch` command. [#3391](https://github.com/pulumi/pulumi/pull/3391)
+
+## 1.4.1 (2019-11-01)
+
 - Adds a **preview** of .NET support for Pulumi. This code is an preview state and is subject
   to change at any point.
 
@@ -24,6 +47,12 @@ CHANGELOG
 
 - `pulumi stack` now renders the stack as a tree view.
   [#3430](https://github.com/pulumi/pulumi/pull/3430)
+
+- Support for lists and maps in config.
+  [#3342](https://github.com/pulumi/pulumi/pull/3342)
+
+- `ResourceProvider#StreamInvoke` implemented, will be the basis for streaming
+  APIs in `pulumi query`. [#3424](https://github.com/pulumi/pulumi/pull/3424)
 
 ## 1.4.0 (2019-10-24)
 
@@ -69,7 +98,7 @@ CHANGELOG
 
 - Support renaming stack projects via `pulumi stack rename`.
   [#3292](https://github.com/pulumi/pulumi/pull/3292)
-  
+
 - Add `helm` to `pulumi/pulumi` Dockerhub container
   [#3294](https://github.com/pulumi/pulumi/pull/3294)